feat: Add mocks validations
