+++ src/mocks/helpers.js
@@ -8,155 +8,17 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const Ajv = require("ajv");
-
-const ajv = new Ajv({ allErrors: true });
-const initAjvErrors = require("ajv-errors");
 const { flatten, compact } = require("lodash");
 
 const tracer = require("../tracer");
+const Mock = require("./Mock");
+const {
+  variantValidationErrors,
+  routeValidationErrors,
+  mockValidationErrors,
+} = require("./validations");
 
 const DEFAULT_ROUTES_HANDLER = "default";
-
-const CLASSES = { Function: Function, RegExp: RegExp };
-
-const HTTP_METHODS = {
-  GET: "get",
-  POST: "post",
-  PATCH: "patch",
-  DELETE: "delete",
-  PUT: "put",
-  OPTIONS: "options",
-  HEAD: "head",
-  TRACE: "trace",
-};
-
-const validHttpMethods = objectKeys(HTTP_METHODS);
-const validHttpMethodsLowerAndUpper = arrayLowerAndUpper(validHttpMethods);
-
-initAjvErrors(ajv, { singleError: true, keepErrors: false });
-
-ajv.addKeyword({
-  keyword: "instanceof",
-  compile: (schema) => (data) => data instanceof CLASSES[schema],
-  errors: false,
-});
-
-const routesSchema = {
-  type: "object",
-  properties: {
-    id: {
-      type: "string",
-      errorMessage: 'Property "id" should be string',
-    },
-    url: {
-      oneOf: [
-        {
-          type: "string",
-        },
-        {
-          instanceof: "RegExp",
-        },
-      ],
-      errorMessage: 'Property "url" should be a string or a RegExp',
-    },
-    method: {
-      oneOf: [
-        {
-          type: "string",
-          enum: validHttpMethodsLowerAndUpper,
-        },
-        {
-          type: "array",
-          uniqueItems: true,
-          items: {
-            type: "string",
-            enum: validHttpMethodsLowerAndUpper,
-          },
-        },
-      ],
-      errorMessage: `Property "method" should be a string or an array with unique items. Allowed values for "method" are "${validHttpMethods.join(
-        ","
-      )}"`,
-    },
-    delay: {
-      type: "integer",
-      minimum: 0,
-      errorMessage: `Property "delay" should be a positive integer`,
-    },
-    variants: {
-      type: "array",
-      items: {
-        type: "object",
-        properties: {
-          id: {
-            type: "string",
-            errorMessage: 'Property "id" should be string in variant ${1#}',
-          },
-          handler: {
-            type: "string",
-            enum: [], // this enum is defined when validator is compiled
-            errorMessage: "", // this error message is defined when validator is compiled
-          },
-          delay: {
-            type: "integer",
-            minimum: 0,
-            errorMessage: 'Property "delay" should be a positive integer in variant ${1#}',
-          },
-        },
-        required: ["id"],
-        errorMessage: {
-          type: "Variant ${0#} should be an object",
-          required: {
-            id: 'Should have a string property "id" in variant ${0#}',
-          },
-        },
-      },
-      errorMessage: `Property "variants" should be an array`,
-    },
-  },
-  required: ["id", "url", "method", "variants"],
-  additionalProperties: {
-    not: true,
-    errorMessage: "Extra property ${0#} is not allowed",
-  },
-  errorMessage: {
-    type: "Should be an object",
-    required: {
-      id: 'Should have a string property "id"',
-      url: 'Should have a property "url"',
-      method: 'Should have a property "method"',
-      variants: 'Should have a property "variants"',
-    },
-  },
-};
-
-let routeValidator;
-
-function compileRouteValidator(routesHandlers) {
-  const supportedRouteHandlersIds = getIds(routesHandlers);
-  routesSchema.properties.variants.items.properties.handler.enum = supportedRouteHandlersIds;
-  routesSchema.properties.variants.items.properties.handler.errorMessage = `Property "handler" should be one of "${supportedRouteHandlersIds.join(
-    ","
-  )}" in variant \${1#}`;
-  routeValidator = ajv.compile(routesSchema);
-}
-
-function toLowerCase(str) {
-  return str.toLowerCase();
-}
-
-function arrayToLowerCase(array) {
-  return array.map(toLowerCase);
-}
-
-function objectKeys(obj) {
-  return Object.keys(obj);
-}
-
-function arrayLowerAndUpper(array) {
-  return array.concat(arrayToLowerCase(array));
-}
 
 function findEqualRouteVariant(routesVariants, routeVariantToFind) {
   return routesVariants.find((routeVariant) => {
@@ -184,12 +46,12 @@ function addMockRoutesVariants(mockRoutesVariants, routesVariantsToAdd) {
 }
 
 function getMockRoutesVariants(mock, mocks, routesVariants, routesVariantsToAdd = []) {
-  const mockRoutesVariants = mock.routesVariants
-    .map((routeId) => {
+  const mockRoutesVariants = compact(
+    mock.routesVariants.map((routeId) => {
       // TODO, add alert if not found
       return routesVariants.find((routeVariant) => routeVariant.variantId === routeId);
     })
-    .filter((route) => !!route);
+  );
   if (mock.from) {
     const from = mocks.find((mockCandidate) => mockCandidate.id === mock.from);
     if (from) {
@@ -285,10 +147,6 @@ function addCustomVariant(variantId, customVariants) {
   return newCustomVariants;
 }
 
-function getIds(objs) {
-  return objs.map((obj) => obj.id);
-}
-
 function hasDelayProperty(obj) {
   return obj.hasOwnProperty("delay");
 }
@@ -299,47 +157,6 @@ function getRouteHandlerDelay(variant, route) {
   }
   if (hasDelayProperty(route)) {
     return route.delay;
-  }
-  return null;
-}
-
-function validationSingleMessage(errors) {
-  return errors
-    .reduce((messages, error) => {
-      if (error.message.length) {
-        messages.push(error.message);
-      }
-      return messages;
-    }, [])
-    .join(". ");
-}
-
-function routeValidationErrors(route) {
-  const isValid = routeValidator(route);
-  if (!isValid) {
-    const idTrace = route && route.id ? `with id "${route.id}" ` : "";
-    return {
-      message: `Route ${idTrace}is invalid: ${validationSingleMessage(routeValidator.errors)}`,
-      errors: routeValidator.errors,
-    };
-  }
-  return null;
-}
-
-function variantValidationErrors(route, variant, Handler) {
-  if (!Handler.validationSchema) {
-    return null;
-  }
-  const variantValidator = ajv.compile(Handler.validationSchema);
-  const isValid = variantValidator(variant);
-  if (!isValid) {
-    const idTrace = variant && variant.id ? `with id "${variant.id}" ` : "";
-    return {
-      message: `Variant ${idTrace}in route with id "${
-        route.id
-      }" is invalid: ${validationSingleMessage(variantValidator.errors)}`,
-      errors: variantValidator.errors,
-    };
   }
   return null;
 }
@@ -412,6 +229,64 @@ function getRouteVariants({ routesDefinitions, addAlert, routeHandlers, core }) 
   );
 }
 
+function getMock({
+  mockDefinition,
+  mockIndex,
+  mocksDefinitions,
+  routeVariants,
+  getGlobalDelay,
+  addAlert,
+  removeAlerts,
+}) {
+  let mock = null;
+  const mockErrors = mockValidationErrors(mockDefinition);
+  const alertScope = `validation:mock:${mockIndex}`;
+  const processAlertScope = `process:mock:${mockIndex}`;
+  removeAlerts(alertScope);
+  removeAlerts(processAlertScope);
+  if (!!mockErrors) {
+    addAlert(alertScope, mockErrors.message);
+    tracer.silly(`Mock validation errors: ${JSON.stringify(mockErrors.errors)}`);
+    return null;
+  }
+  try {
+    mock = new Mock({
+      id: mockDefinition.id,
+      routesVariants: getMockRoutesVariants(mockDefinition, mocksDefinitions, routeVariants),
+      getDelay: getGlobalDelay,
+    });
+  } catch (error) {
+    addAlert(processAlertScope, error.message);
+    tracer.error(`Error processing mock: ${error.message}`);
+  }
+  return mock;
+}
+
+function getMocks({ mocksDefinitions, addAlert, removeAlerts, routeVariants, getGlobalDelay }) {
+  let errorsProcessing = 0;
+  const mocks = compact(
+    mocksDefinitions.map((mockDefinition, index) => {
+      const mock = getMock({
+        mockDefinition,
+        mockIndex: index,
+        mocksDefinitions,
+        routeVariants,
+        getGlobalDelay,
+        addAlert,
+        removeAlerts,
+      });
+      if (!mock) {
+        errorsProcessing++;
+      }
+      return mock;
+    })
+  );
+  if (errorsProcessing > 0) {
+    addAlert("process:mocks", `${errorsProcessing} errors found while loading mocks`);
+  }
+  return mocks;
+}
+
 module.exports = {
   getMockRoutesVariants,
   getVariantId,
@@ -419,14 +294,10 @@ module.exports = {
   getPlainRoutes,
   getPlainRoutesVariants,
   addCustomVariant,
-  getIds,
   getRouteHandlerDelay,
-  routeValidationErrors,
-  variantValidationErrors,
-  HTTP_METHODS,
-  compileRouteValidator,
   getVariantHandler,
   getRouteVariants,
   hasDelayProperty,
-  validationSingleMessage,
+  getMock,
+  getMocks,
 };
