+++ src/mocks/validations.js
@@ -1,0 +1,256 @@
+/*
+Copyright 2021 Javier Brea
+
+Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+*/
+
+const Ajv = require("ajv");
+
+const ajv = new Ajv({ allErrors: true });
+const initAjvErrors = require("ajv-errors");
+
+const HTTP_METHODS = {
+  GET: "get",
+  POST: "post",
+  PATCH: "patch",
+  DELETE: "delete",
+  PUT: "put",
+  OPTIONS: "options",
+  HEAD: "head",
+  TRACE: "trace",
+};
+
+const CLASSES = { Function: Function, RegExp: RegExp };
+
+const validHttpMethods = objectKeys(HTTP_METHODS);
+const validHttpMethodsLowerAndUpper = arrayLowerAndUpper(validHttpMethods);
+
+initAjvErrors(ajv, { singleError: true, keepErrors: false });
+
+ajv.addKeyword({
+  keyword: "instanceof",
+  compile: (schema) => (data) => data instanceof CLASSES[schema],
+  errors: false,
+});
+
+const mocksSchema = {
+  type: "object",
+  properties: {
+    id: {
+      type: "string",
+      errorMessage: 'Property "id" should be string',
+    },
+    from: {
+      type: "string",
+      errorMessage: 'Property "from" should be string',
+    },
+    routesVariants: {
+      type: "array",
+      uniqueItems: true,
+      items: {
+        type: "string",
+      },
+      errorMessage: 'Property "routesVariants" should be an array of strings with unique items',
+    },
+  },
+  required: ["id", "routesVariants"],
+  additionalProperties: {
+    not: true,
+    errorMessage: "Extra property ${0#} is not allowed",
+  },
+  errorMessage: {
+    type: "Should be an object",
+    required: {
+      id: 'Should have a string property "id"',
+      routesVariants: 'Should have a property "routesVariants"',
+    },
+  },
+};
+
+const routesSchema = {
+  type: "object",
+  properties: {
+    id: {
+      type: "string",
+      errorMessage: 'Property "id" should be string',
+    },
+    url: {
+      oneOf: [
+        {
+          type: "string",
+        },
+        {
+          instanceof: "RegExp",
+        },
+      ],
+      errorMessage: 'Property "url" should be a string or a RegExp',
+    },
+    method: {
+      oneOf: [
+        {
+          type: "string",
+          enum: validHttpMethodsLowerAndUpper,
+        },
+        {
+          type: "array",
+          uniqueItems: true,
+          items: {
+            type: "string",
+            enum: validHttpMethodsLowerAndUpper,
+          },
+        },
+      ],
+      errorMessage: `Property "method" should be a string or an array with unique items. Allowed values for "method" are "${validHttpMethods.join(
+        ","
+      )}"`,
+    },
+    delay: {
+      type: "integer",
+      minimum: 0,
+      errorMessage: `Property "delay" should be a positive integer`,
+    },
+    variants: {
+      type: "array",
+      items: {
+        type: "object",
+        properties: {
+          id: {
+            type: "string",
+            errorMessage: 'Property "id" should be string in variant ${1#}',
+          },
+          handler: {
+            type: "string",
+            enum: [], // this enum is defined when validator is compiled
+            errorMessage: "", // this error message is defined when validator is compiled
+          },
+          delay: {
+            type: "integer",
+            minimum: 0,
+            errorMessage: 'Property "delay" should be a positive integer in variant ${1#}',
+          },
+        },
+        required: ["id"],
+        errorMessage: {
+          type: "Variant ${0#} should be an object",
+          required: {
+            id: 'Should have a string property "id" in variant ${0#}',
+          },
+        },
+      },
+      errorMessage: `Property "variants" should be an array`,
+    },
+  },
+  required: ["id", "url", "method", "variants"],
+  additionalProperties: {
+    not: true,
+    errorMessage: "Extra property ${0#} is not allowed",
+  },
+  errorMessage: {
+    type: "Should be an object",
+    required: {
+      id: 'Should have a string property "id"',
+      url: 'Should have a property "url"',
+      method: 'Should have a property "method"',
+      variants: 'Should have a property "variants"',
+    },
+  },
+};
+
+const mockValidator = ajv.compile(mocksSchema);
+let routeValidator;
+
+function getIds(objs) {
+  return objs.map((obj) => obj.id);
+}
+
+function compileRouteValidator(routesHandlers) {
+  const supportedRouteHandlersIds = getIds(routesHandlers);
+  routesSchema.properties.variants.items.properties.handler.enum = supportedRouteHandlersIds;
+  routesSchema.properties.variants.items.properties.handler.errorMessage = `Property "handler" should be one of "${supportedRouteHandlersIds.join(
+    ","
+  )}" in variant \${1#}`;
+  routeValidator = ajv.compile(routesSchema);
+}
+
+function toLowerCase(str) {
+  return str.toLowerCase();
+}
+
+function arrayToLowerCase(array) {
+  return array.map(toLowerCase);
+}
+
+function objectKeys(obj) {
+  return Object.keys(obj);
+}
+
+function arrayLowerAndUpper(array) {
+  return array.concat(arrayToLowerCase(array));
+}
+
+function validationSingleMessage(errors) {
+  return errors
+    .reduce((messages, error) => {
+      if (error.message.length) {
+        messages.push(error.message);
+      }
+      return messages;
+    }, [])
+    .join(". ");
+}
+
+function mockValidationErrors(mock) {
+  const isValid = mockValidator(mock);
+  if (!isValid) {
+    const idTrace = mock && mock.id ? `with id "${mock.id}" ` : "";
+    return {
+      message: `Mock ${idTrace}is invalid: ${validationSingleMessage(mockValidator.errors)}`,
+      errors: mockValidator.errors,
+    };
+  }
+  return null;
+}
+
+function routeValidationErrors(route) {
+  const isValid = routeValidator(route);
+  if (!isValid) {
+    const idTrace = route && route.id ? `with id "${route.id}" ` : "";
+    return {
+      message: `Route ${idTrace}is invalid: ${validationSingleMessage(routeValidator.errors)}`,
+      errors: routeValidator.errors,
+    };
+  }
+  return null;
+}
+
+function variantValidationErrors(route, variant, Handler) {
+  if (!Handler.validationSchema) {
+    return null;
+  }
+  const variantValidator = ajv.compile(Handler.validationSchema);
+  const isValid = variantValidator(variant);
+  if (!isValid) {
+    const idTrace = variant && variant.id ? `with id "${variant.id}" ` : "";
+    return {
+      message: `Variant ${idTrace}in route with id "${
+        route.id
+      }" is invalid: ${validationSingleMessage(variantValidator.errors)}`,
+      errors: variantValidator.errors,
+    };
+  }
+  return null;
+}
+
+module.exports = {
+  HTTP_METHODS,
+  getIds,
+  mockValidationErrors,
+  routeValidationErrors,
+  variantValidationErrors,
+  compileRouteValidator,
+  validationSingleMessage,
+};
